# openapi.yaml
openapi: 3.0.3
info:
  title: Push Balancer API
  description: |
    ML-gestützte API für BILD Push-Notification-Optimierung.

    Der Push Balancer analysiert historische Push-Daten, prognostiziert Opening Rates (OR)
    via Gradient Boosted Trees und Stacking Ensemble (LightGBM + XGBoost + CatBoost → Ridge),
    und liefert tagesaktuelle Empfehlungen für Redaktion und Sport.

    **Safety**: Alle Vorhersagen sind ausschließlich beratend (ADVISORY_ONLY). Das System
    darf niemals autonom Push-Benachrichtigungen versenden. Jede Prediction enthält
    `advisoryOnly: true` und `actionAllowed: false`.
  version: 2.4.0
  contact:
    name: BILD Digital Produkt
    email: push-balancer@axelspringer.com
servers:
  - url: http://localhost:8050
    description: Lokaler Entwicklungsserver
  - url: https://push-balancer.onrender.com
    description: Render Cloud Deployment
...
paths:
...
  /api/gbrt/status:
    get:
      operationId: getGbrtStatus
      tags: [gbrt]
      summary: GBRT-Modell-Status abrufen
      description: |
        Umfassender Status des Gradient Boosted Regression Trees Modells:
        Anzahl Bäume, Feature Importances, Kalibrierungs-Bins, Trainings-Metriken,
        Error-Analyse nach Kategorie und Stunde sowie letzte 20 Predictions aus
        dem Prediction-Log.
      responses:
        "200":
          description: GBRT-Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GbrtStatusResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/gbrt/model.json:
    get:
      operationId: getGbrtModelJson
      tags: [gbrt]
      summary: GBRT-Modell für Client-Side-Evaluation exportieren
      description: |
        Exportiert das serialisierte GBRT-Modell als JSON für Client-Side-Inference
        im Browser (JavaScript-Implementierung). Das Modell wird 5 Minuten gecacht
        (`Cache-Control: max-age=300`).
      responses:
        "200":
          description: Serialisiertes GBRT-Modell als JSON
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Serialisiertes Baumstruktur-JSON (baumspezifisches Format)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/gbrt/predict:
    get:
      operationId: getGbrtPredict
      tags: [gbrt]
      summary: Einzelne OR-Prognose (GBRT)
      description: |
        Berechnet eine OR-Prognose via dem serverseitigen GBRT-Modell. Liefert
        neben der Prognose auch Quantile (Q10/Q90) und eine Konfidenzangabe.
        Falls das Modell noch nicht trainiert ist, wird ein `error`-Feld zurückgegeben.
      parameters:
        - name: title
          in: query
          required: false
          description: Push-Titelzeile
          schema:
            type: string
            example: "Sensation im Bundestag: Das ist der neue Kanzler"
        - name: cat
          in: query
          required: false
          description: Artikel-Kategorie
          schema:
            type: string
            default: News
            example: Politik
        - name: hour
          in: query
          required: false
          description: Stunde des geplanten Versands (0–23), Standard ist aktuelle Stunde
          schema:
            type: integer
            minimum: 0
            maximum: 23
        - name: eilmeldung
          in: query
          required: false
          description: "1 wenn Eilmeldung, sonst 0"
          schema:
            type: integer
            enum: [0, 1]
            default: 0
        - name: plus
          in: query
          required: false
          description: "1 wenn BILD-Plus-Artikel, sonst 0"
          schema:
            type: integer
            enum: [0, 1]
            default: 0
      responses:
        "200":
          description: GBRT-Prognose
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GbrtPrediction"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/gbrt/retrain:
    post:
      operationId: triggerGbrtRetrain
      tags: [gbrt]
      summary: GBRT-Modell neu trainieren (manuell)
      description: |
        Startet asynchron ein Neu-Training des GBRT-Modells im Hintergrund.
        Das Training läuft via Early-Stopping und Kalibrierung.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Training gestartet
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: gbrt_training_started
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/gbrt/force-promote:
    post:
      operationId: forcePromoteGbrt
      tags: [gbrt]
      summary: Gespeichertes GBRT-Modell als Champion laden
      description: |
        Lädt das zuletzt in `gbrt_model.json` gespeicherte Modell als aktiven Champion
        in den Speicher. Deaktiviert einen laufenden A/B-Test. Nützlich nach manuellem
        Modell-Austausch oder nach einem Rollback.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Modell erfolgreich promoted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: force_promoted
                  experimentId:
                    type: string
                    example: exp_20260401_143022
                  testMae:
                    type: number
                    format: float
                    example: 1.42
                  modelType:
                    type: string
                    example: gbrt_ensemble
                  nFeatures:
                    type: integer
                    example: 38
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Tagesplan ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details für HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI-Referenz, die den Fehlertyp identifiziert
          example: "https://push-balancer.example.com/problems/validation-error"
        title:
          type: string
          description: Kurze, menschenlesbare Fehlerbeschreibung
          example: "Validation Error"
        status:
          type: integer
          description: HTTP-Statuscode
          example: 400
        detail:
          type: string
          description: Ausführlichere Fehlerbeschreibung
          example: "Das Feld 'pushId' darf nicht leer sein"
        instance:
          type: string
          format: uri
          description: URI-Referenz des konkreten Fehler-Vorkommens
          example: "/api/prediction-feedback"

    # ── Health ──────────────────────────────────────────────────────────────
...
    MlMetrics:
      type: object
      properties:
        mae:
          type: number
          format: float
          description: Mean Absolute Error in Prozentpunkten
          example: 1.42
        rmse:
          type: number
          format: float
          description: Root Mean Squared Error in Prozentpunkten
          example: 2.01
        r2:
          type: number
          format: float
          description: R²-Score (Bestimmtheitsmaß)
          example: 0.71
        trainN:
          type: integer
          description: Anzahl Trainings-Samples
          example: 2800
        valN:
          type: integer
          description: Anzahl Validierungs-Samples
          example: 600
        testN:
          type: integer
          description: Anzahl Test-Samples
          example: 400
    FeatureImportanceEntry:
      type: object
      properties:
        name:
          type: string
          description: Feature-Name
          example: hour_sin
        importance:
          type: number
          format: float
          description: Relative Wichtigkeit (0–1)
          example: 0.187
...
    GbrtPrediction:
      type: object
      properties:
        predicted:
          type: number
          format: float
          description: Prognostizierte Opening Rate in Prozent
          example: 6.92
        q10:
          type: number
          format: float
          example: 4.1
        q90:
          type: number
          format: float
          example: 10.3
        confidence:
          type: number
          format: float
          example: 0.71
        error:
          type: string
          nullable: true
          description: Fehlermeldung falls Modell nicht trainiert
          example: null
    GbrtStatusResponse:
      type: object
      properties:
        loaded:
          type: boolean
        trained:
          type: boolean
        message:
          type: string
          description: Fehlermeldung falls nicht trainiert
          example: null
        model:
          type: object
          properties:
            type:
              type: string
              example: GBRT
            nTrees:
              type: integer
              example: 180
            nFeatures:
              type: integer
              example: 38
            nPushes:
              type: integer
              example: 3820
            mae:
              type: number
              format: float
              example: 1.42
            r2:
              type: number
              format: float
              example: 0.74
            trainedAt:
              type: number
              example: 1743500000
            modelType:
              type: string
              example: gbrt_ensemble
        metrics:
          $ref: "#/components/schemas/MlMetrics"
        featureImportance:
          type: array
          items:
            $ref: "#/components/schemas/FeatureImportanceEntry"
        calibration:
          type: object
          properties:
            bins:
              type: array
              items:
                type: object
                properties:
                  predicted:
                    type: number
                  actual:
                    type: number
                  count:
                    type: integer
            calibrationError:
              type: number
              format: float
        errorAnalysis:
          type: object
          properties:
            byCategory:
              type: object
              additionalProperties:
                type: object
                properties:
                  mae:
                    type: number
                  bias:
                    type: number
                  n:
                    type: integer
            byHour:
              type: object
              additionalProperties:
                type: object
                properties:
                  mae:
                    type: number
                  bias:
                    type: number
                  n:
                    type: integer
        recentPredictions:
          type: array
          items:
            $ref: "#/components/schemas/RecentPrediction"
    RecentPrediction:
      type: object
      properties:
        title:
          type: string
        predicted:
          type: number
          format: float
        actual:
          type: number
          format: float
          nullable: true
        confidence:
          type: number
          format: float
        q10:
          type: number
          format: float
        q90:
          type: number
          format: float

    # ── Tagesplan ───────────────────────────────────────────────────────────
...
  responses:
...
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/not-found"
            title: "Not Found"
            status: 404
            detail: "GBRT-Modell nicht vorhanden"
...
    InternalServerError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://push-balancer.example.com/problems/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "ML status error: model not initialized"
...